test: Add coverage for Transform _flush async case.
This is new failing  test related to the _flush method of transform
streams.  It adds coverage for the case when _flush makes an async call.

Previously there was no coverage this functionality.

The 'finish' event is documented to fire with flushing is done: When
"all data has been flushed to the underlying system".

However, it appears be fired too early in the case of Transform streams
which make an async call. process.nextTick is used here but I ran into
this in a real-world case of making a call to a Mongo database.

This new failing test illustrates the apparent bug. When the code is
_flush is synchronous, there's no problem.

It fails with both 0.10.x and the latest from the 0.11 branch as well.